backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
openpyxl==3.1.2
reportlab==4.0.7
pytz==2024.1
